+++ Common/Common.h
@@ -4,7 +4,7 @@
 
 namespace Common
 {
-    double mod(double vVal, double vMax);
+    float mod(float vVal, float vMax);
     std::vector<char> readFileAsChar(std::filesystem::path vFilePath);
 
     template <typename T>
